Move prop gathering to routing component.
