+++ src/components/organizations/list/organization_row.js
